dse-k8s-eqiad: substantially increase job sidecar controller CPU resources
We observed that it takes quite a while for the task pods to transition from the
`2/3 Unready` state to the `0/3 Terminating` state, and when looking at
the pod resources dashboard, we see a lot of CPU throttling happening
for the sidecar controller while the dumps are running.

Let's see if increasing its CPU resources will cause the pods to be
deleted faster.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T392995
Change-Id: Ie0f7dc75126a367812621583185aa7c72394ad82