Add an ssh known_hosts configmap to the mediawiki-dumps-legacy namespace
We need to be able to use rsync over ssh from the mediawiki-dumps-legacy
namespace and the clouddumps servers.

It order to validate that the hosts are correct, we take the known_hosts
entries from https://config-master.wikimedia.org/known_hosts and put the
clouddumps100[1-2] keys into a configmap that will be mounted by the
sync pods.

Bug: T390738
Change-Id: Id1c11bf664dea4e40d083f97fe100205beb5a115