airflow: prevent resource name collisions when multiple releases are installed in the same namespace
This should not induce any change for the production-bearing airflow
instances, as we're simply moving the volume/volume_mounts data blocks
from values.yaml to some templates, to add the ability to insert the
release name in the resource name prefix.

For example, the airflow-scheduler Deployment would become
airflow-devenv-brouberol-scheduler, if the release name is
`devenv-brouberol`.

I'm doing things this way to avoid changing the name of resources such
as volumes, that do not tolerate a rename without inducing data loss, as
a "rename" from A to B with helm is really just a delete A + create B.

There is however one specific case where this patch will induce a change
for non-devenvs releases: the (currently named) `envoy` Service. We need
that service name to be prefixed with the release name as well for
devenvs, however we've already hardcoded the `envoy` service name in
airflow-dags.
I've currently duplicated the service to both `envoy` and
`<release-prefix>-envoy` to give us the time to adjust the airflow-dags
code, aka the service name in the `k8s_proof_url` function.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T393999
Change-Id: Ib2c0b4abf8760cfeefedfc0a9516f8d006d7fc5e