calico: Allow to override the MTU via values files
Why:
Per T352956, to adopt Liberica and the new networking setup, we need
IPIP encapsulation and a way to inform clients of our TCP MSS size,
which due to encapsulation needs to be different than the default of
1480. The TCP handshake does use the TCP MSS option exactly for that reason, but
the pod needs to know the MTU it has. Calico does have the veth_mtu
option in the configmap exactly for this reason.

What:
Allow to set the veth_mtu configuration via values.yaml

Bug: T352956
Change-Id: Ibd98710f0af5fa0c1ebfcb516ee852974d6f7707