airflow: rely on krenew instead of 'airflow kerberos' to renew the kerberos ticket
In development environments, we cannot rely on a keytab, as the kerberos token is associated
with the personal identity of the developer, instead of a service account. Due to the lack of
keytab, we can't rely on the `airflow kerberos` command (which assumes a keytab is present),
so we rely on the `krenew` command, to simpluy renew the ticket contained in the krb5 cache.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T393999
Bug: T394293
Change-Id: I086dcfa8ee799414d3b42e2143124b962f5ac1aa