mw-videoscaler: drop job concurrency
Currently we very rarely see a backlog for video encoding jobs in
general, even though we fully expect to see one. Having this level of
concurrency is probably over generous, and when we have 5+ sync-worlds
running one after another, we end up with $scaps*$concurrency workers
used, which causes pages.

Change-Id: I5d32190f7d85036cad6d33f6d20ba58ef297dfe0