datahub: increase resources accross the board
We woke up to datahub OOMing accross the board, causing instabilities in
airflow as tasks have a direct dependency on datahub, that we need to
investigate.

In an "increase resources first, ask questions later" we bumped
resources to all components until they concedeed that they should maybe
do as they're told.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: Ic1f2002e312e4d0134f52c5e0d50893be8f1a38a