Remove wmf.volumes from all charts
Why:
wmf.volumes is an older mechanism that hasn't been kept up to date since
the move to sextant. It isn't really used, except as a thin wrapper
around sextant modules. However, there is some drift between charts
which poses the question of where this can go if left as is.

What:
* Remove all template files and inclusions
* Amend the charts that actually use the functionality
* Add a test for volumes used by shellbox
* Skip 1 chart, aqs-http-gateway as CI is failing for geo-analytics
  service. Break down the problem to a smaller chunk

Change-Id: Ie57bd7531e329c64343accce90ac9d03308b2f21