changeprop: Decrease reenqueue_delay for Getting Started notif job
This alters the config for notificationGetStartedJob by
decreasing the reenqueue delay to 30 minutes (from 72 hours).
This is because Product requested sending the notification
sooner to certain users, which is not possible with
the previous reenqueue delay.

We did a research spike on the effect of this change,
but we were unable to reach conclusions. As such, this
is an experiment -- we will want to monitor the effects
(hopefully none). Fortunately, the job has a dedicated
changeprop queue, so changes _should_ only impact this
particular job.

Bug: T394958
Change-Id: Ib36e2935eca16352cb102c411431f787678b5d69