kserve-inference: set seccomp defaults in the chart
Up to now we have been running with the custom seccomp settings
only in ml-serve-codfw and ml-staging. This change moves the special
settings to the kserve-inference chart, to remove unnecessary code
duplication and DRY the configs.

The consequence of the change is that we'll be able to rollout
seccomp settings also to ml-serve-eqiad.

Bug: T369493
Change-Id: I06237c90adebd633cf8280bb618806d959d8b9d6