jobqueue: Set the host header in all jobs
Why:
We want to emit the information required by the MediaWiki single
version HTTP routing implementation in our produced jobs, namely the
HTTP host header referring to the domain that the job is for. This
information already exists in the job events per

https://schema.wikimedia.org/repositories//primary/jsonschema/mediawiki/job/1.0.0

and a kafkacat of various jobs in the meta object of the job.

What:
Set the host header to message.meta.domain

Bug: T395451
Change-Id: Ia20ef2c142f82346a09e4dd023d0791aadc6dd90