mw-jobrunner: VirtualHost priority to 0
Why:
Despite this being the only website around in the Apache configuration,
we also have the nonexistent one by default in our setup. That one ends
up with a virtualhost file entry of 00-nonexistent. Having 0 for the
priority here as well, ensures that all requests go to the mw-jobrunner
one, even if they don't have the proper Host header. This is crucial for
the Single Version MediaWiki HTTP routing implementation

What:
Switch priority to 0

Bug: T395451
Change-Id: I82ff75650f701d79cb34b834e8a1200d03674208