Airflow: Add local settings to enable the xcom_sidecar functionality
When attempting to use the `do_xcom_push=True` feature of Airflow with a
task using the KubernetesPodOperator, the default behaviour of Airflow
is to attempt to create an xcom_sidecar container within the task pod.

As per:
https://airflow.apache.org/docs/apache-airflow-providers-cncf-kubernetes/stable/operators.html#how-does-xcom-work

The default settings try to launch an alpine image with resource
settings that are lower than the minimum values allowed by our cluster.
In addition to that, the required security context for all of our pods
is missing.

This patch configures the xcom_sidecar pod defaults in the
local_settings.py configuration file, so that this functionality will be
enabled.

Bug: T388378
Bug: T396197
Change-Id: Id91de6c18fa593d9fc10060100074cab8b58673f