rest-gateway: per-route statistics option, enable for lists and wikifeeds
In order to get more insight about indivudual route behaviour,
performance and availability, add the option to use a stat prefix for
any route. Envoy docs warn that this will add 1KiB memory use per
instance, but we don't have that many routes at present, so enabling
it by default in future shouldn't be that big of a hit. For the
short-term, enable it for the reading lists API and all of wikifeeds

Example output when testing against the devel instance available at
https://phabricator.wikimedia.org/P77195

Change-Id: I546797941eee1bef07d4b3f8f78f0e826758d359