(api|rest)-gateway: remove envoyproxy annotation, scrape all ports
When the gateway was first built, scraping multiple ports was a more
difficult affair and the envoyproxy.io annotation was abused to allow
us to scrape the statsd gateway for the ratelimit service. This is no
longer required, and we can simply scrape all containerPorts as
normal. Currently we are missing some metrics from the gateway's envoy
instance as using the envoyproxy.io annotation only scrapes particular
metrics, and we need more than those.

Change-Id: I1d2c0e71cd9ffb0ca4e18f859788ed026521c998