Bump up the CPU and RAM resources for airflow related namespaces
With the increasing workload moving to Airflow recently, including
dumps_v1 and refinery, we need to increase the resource quotas for the
related namespaces.

These are:
* airflow-main
* airflow-test-k8s
* mediawiki-dumps-legacy

We are over-provisioning a little, but at the moment we are throttling
our Airflow pipelines unnecessarily, so we will continue to monitor.

In time, the dumps_v1 DAGs will move the airflow-main instance, so we
may be able to reduce the resource quotas on that.

Bug: T388378
Change-Id: I5623471e91be352ef07340a6f61da2b9fd7b813f