airflow: tweak the environment variable exposing whether we're in dev/prod
Having having talked with mforns, we realized that we already had code
in airflow-dags that relies on the AIRLFOW_ENVIRONMENT_NAME env var,
with similar intent and a slightly different meaning:
the value would also define airflow default task variables, and the
values come from a dictionary with `wmf` and `dev_wmf` keys.

We can piggyback on that existing behavior to define a function that
returns whether we run in a devenv or not (irrespective of whether it
was created with the legacy or new method).

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T394297
Change-Id: I63c9b0c8cb50ad9a79339264f4e29b03786704c4