changeprop beta: Decrease reenqueue_delay for Getting Started notif job
This mimics the production change, which was done in
Ib36e2935eca16352cb102c411431f787678b5d69.

This alters the config for notificationGetStartedJob by
decreasing the reenqueue delay to 30 minutes (from 72 hours).
This is because Product requested sending the notification
sooner to certain users, which is not possible with
the previous reenqueue delay.

Bug: T394958
Change-Id: I7ef1560f7852d7dfef19e3fe51ff1e5ba2510889