calico: Switch default-deny to using services instead of ports
Why:
e68abcdaf switched the port that CoreDNS pods listen on, for good reason
to 5353. This has been deployed to staging clusters already. The default
deny rule just allowlists DNS but the way it does targets the 53 port
(and by virtue of not defining anything else, default to denying
everything else).

That change broke wikifunctions, but not any other namespace. The reason
for the latter is that  the allow-pod-to-pod rule that allows egress
from all pods exists. However, that rule explicitly excludes
wikifunctions because of the tigher rules required there

Instead of adding the 5353 port, let's switch to something that targets
the service.

What:
Switch from endpoint selector and ports to just naming out the service
we want to allow

Bug: T397341
Change-Id: Icdd50a0a52c09f160fbfef074d3efa85018aa757