mobileapps: set resource limits for statsd exporter
We have removed default limits for the mobileapps namespace in order
to allow us to rapidly scale up in an emergency and also experiment
with the OOMKill issues we are seeing in prod. Set a limit for the
statsd pods as the API complains about a lack of limit here.

Bug: T397703
Change-Id: I5fd27dca44a949e23d656946781f407398433d62