dse-k8s: Configure limitranges for mediawiki-dumps-legacy
We need to be able to run pods with >16 GB of RAM in order to run the
cirrussearch dumps, as well as some of the larger jobs to recombine
multistream versions of the sql/xml dumps.

In order to support these, increase the maximum pod/container limits in
the mediawiki-dumps-legacy namespace.

Additionally, we wish to allow very small containers and pods,
specifically to support the xcom_sidecar functionality of the Airflow
KubernetesPodOperator.

Bug: T394389
Change-Id: I408829bd313d10d455e2cf919f59b288a51db80c