mobileapps: bump memory limits without scaling down
Significantly increase memory limits to attempt to deal with latency
and processing times in changeprop. Previously we scaled down
alongside these increases. This time, we don't scale down and the
limit is slightly over the previous limit bump. This is probably not a
reasonable value to keep forever, but is to experiment with whether we
can realistically address this issue with resourcing at all.

Bug: T397750
Change-Id: Ic5def7e507dc0fd1fe6a2d9b0f875ecf7853f3ba