mw-parsoid: Scale down replicas
Following the migration of parsoid endpoints out of restbase and to
rest-gateway, the backing for these calls is mw-api-int and not
mw-parsoid. As such, we can dramatically scale down the mw-parsoid
replica count.

It is currently handling about 40 ops/s, for a 12.5% fpm worker
saturation, so we can safely go down to 25 replicas.

Change-Id: I5d9f65d60424372c6a5fdce485386f627f7decce