Airflow-main: Increase parallelism and related values
We wish to increase the parallelism of the airflow-main instance, since
our number of task instances created by our DAGs is increasing an we
need to ensure a high throughput.

Consequently, we increase the parallelism value from 128 to 256, as well
as the max_tis_per_query and worker_pods_creation_batch_size and the
number of vCPU cores for the scheduler process.

Bug: T398164
Change-Id: Ic3429d41305b8553998aba12db9891ea810b2329