api-gateway: use more recent ratelimit image
We have been using a very old version of the ratelimit service in the
gateway for a long time. This image is buster-based and needs to be
replaced as is, and the newer version will have features we will need
in the coming quarters. Only use the new version on staging for now.

Bug: T388804
Change-Id: Id7781b3135991aae54a98596b27d0e78362ad54b