Bump resources and shared buffers config for postgresql-airflow-main
We have seen occurrences whereby the postgresql servers behind
airflow-main are approaching their limits in terms of max_connections
and CPU allocation.

In response to this, we want to increase the CPU and memory allocation,
plus we can start to tune the Postgres configuration itself to match the
resources more closely. We will start by defining the shared_buffers to
be 20% of the available RAM in the container.

This is based on the guidance here:
https://www.postgresql.org/docs/current/runtime-config-resource.html#GUC-SHARED-BUFFERS

Bug: T398421
Change-Id: If27b5f73ea524297890096f5bf0f56b05e52526e