airflow: enable Kerberos security in devenvs
We have tracked down a bug in devenvs caused by `conf.get("core", "security")`
returning `''` instead of `kerberos`, causing the Hive metastore client
to fail to receive data from the metastore.

As we're not using a Kerberos auth backend in devenvs, this does not
prevent us from using Basic Auth login, and having an open API in dev.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T398683
Change-Id: I57c767bcb9b815485e900ee26863e63961bc1240