Tweak the limitrange for the dse-k8s-eqiad/spark-history namespace
We want to allow the spark-history service to use a maximum of 16 CPU
cores, but the limitranges were set such that a pod's maximum CPU was
16.

In our case, this prevented spark-history from using the value we
wanted, because there was no excess for the envoyproxy container within
the pod.

This change increases the maximum CPU to 17000m for the spark-hostory
namespace.

Bug: T396617
Change-Id: I5d6a998f72e4d22cd5071c690adbdc8901ae2cfb