thumbor: Lower thumbor_workers, more memory
Thumbor tends to not use as many workers as we define, but the ones that
are used OOM frequently, or simply fail because of the memory limit in
the imagemagick policy file.
Lower the number of thumbor_workers per pod from 7 to 4, and raise their
memory limit from 3 to 7Gi.

Bug: T392348
Change-Id: I0441d2eba8e401acbe69bbc209df4138733e4e2b