thumbor: change haproxy load balancing algorithm to leastconn
thumbor has long displayed an issue where it rarely hits the
higher-numbered worker containers while putting excess load on the
lower numbered pods. This is because the `first` lb algo will ensure
that the first pod hits maxconns and then move on to the next. This is
very inefficient. Use least connections instead.

Bug: T392348
Change-Id: I86befee0880e94bbe7e8cf389a48716fb63a23c8