BGPPeers nodeSelector: remove old codfw rows, add future eqiad pods
The removal of the old codfw rows goes with:
    Ib7e7e3afd17bee452833f57d7cdedb15ce5686ed

The addition of the "pod-ab/cd" will take effert temporarily when
the new switches (and network topology) are setup in eqiad.
Then we will be able to remove them (as well as the row wide zones) once
all the k8s nodes are migrated to their per rack vlans.
Leaving only the k8s running on Ganeti to peer with the core routers.

I'm not touching the single rack zones "row-e1", even though their name
is not ideal :)

Bug: T333948
Change-Id: Ie67c2930d9b22ebd2d076c3ea8c099d0bb4a35f1