Allow the Airflow webserver to support long requests
We have, on occasion, seen Airflow close connections when we have asked
it to clear many tasks. We would therefore like to raise the limit of
how long a requests may take to return from the gunicorn web server.

In addition to this, we need to increase the amount of time that
envoyproxy will wait for a response.

We have increased this limit from 60 seconds to 300 seconds.

Bug: T400493
Change-Id: I94bab4155bf1bc106f198db1ca78450e8a9e3e07