airflow: avoid configmap name collision when spawning multiple devenvs
@xcollazo realized that spawning multiple devenvs with --dags-folder main
failed, due to a ConfigMap name collision.

This was due to the fact that we were passing `.Root` to the template in
charge of rendering the configmap name, within a `range` loop. In these
loops, `.` becomes the current loop context, meaning that `.devenv` was
evaluated to `nil`, which is falsy, and thus the configmap name was not
using the devenv release name.

I have confirmed this to be working by deploying 2 devenvs, using
```
airflow-devenv create \
    --name dev-brouberol{1,2} \
    --branch main \
    --dags-folder main \
    --charts-folder ~/deployment-charts/charts/
```

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I1de160ab95bdec9014e6204dc1c9cfc0df6ec8e4