mesh: networkpolicy: also allow egress to 4318 (http)
Why:
When Envoy's tracing support is enabled, we allow egress to 4317,
the grpc port of the otel collector.  This is sufficient for Envoy.

But services which opt to enable tracing will often want to also send
their own spans to the otel collector, and may want to use http
instead of grpc.  Let's allow this.

Since the otelcol runs as a DaemonSet of pods on k8s,
this was already allowed by the default pod-to-pod rule --
*except* for WikiFunctions, which is denied pod-to-pod.

Bug: T344177
Bug: T390753
Bug: T375922
Change-Id: I42fd50475f7e4101328661e476b2959a70ce020a