mediawiki: Add support for mounting a custom dblist
Accept the contents of a dblist in Values.mwscript.dblist_contents and
map it in as /srv/mediawiki/dblists/mwscript.dblist so it can be used
with foreachwikiindblist.

Rename the existing Values.mwscript.dblist to Values.mwscript.dbexpr,
and the annotation too. ("dbexpr" is more accurate anyway, and "dblist"
could awkwardly be a good name for either thing, so let's use it for
neither).

Bug: T401737
Change-Id: I98733e24c28889f9672ff23fa6dd37324d7e1874