airflow: make it possible to inject custom files in secrets
The datahub ingestion files we have in airflow-dags do not contain any
secret, as they rely on Kerberos authentication (or no auth at all).
As we'd like to integrate Datahub and Superset, we have defined a
`datahub` user in the Superset DB, with a password.

We add a feature to the airflow chart, through which we make it possible
to render custom secret files under ${AIRFLOW_HOME}/secrets, thus
allowing us to define the a custom datahub ingestion config file with a
password in the private puppet repo, render it in the Secret resource via helmfile,
and have our DAG simply copy the file to its SkeinOperator application directory.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T309622
Change-Id: I467df0a6ae0a5be81f55e694a58a3bed97225c24