rest-gateway: allow route definition to reuse clusters
We need more flexibility in how we define routes. We want to route
rest.php to mw-api-int (for now - should be -ext in future), but we
don't want to stop the restbase compatibility layers we have in place
for ex-restbase APIs. Introduce the "cluster" parameter which lets a
set of routes have their own stat namespace and matching behaviours
but reuse the cluster definition.

Move all rest.php routes to their own section - disable the
x-restbase-compat header for that route set (which leads to incorrect
API responses if set), and remove the restbase header set which will
surely need to be changed in future and will not fit with the
requirements of rest.php.

Bug: T400132
Change-Id: I057e520e1129e6e5c60852fef8808b26056e4bca