flink chart: Add a comment label
We need a way to positively identify, from an orchestration
script, that a backfill release from the cirrus-streaming-updater
is a specific invocation of the backfill release.  Facilitate
this by adding a `comment` label that can be provided. This label
can optionally contain the script name and a unique id, allowing
for specific mapping of the deployment back to an orchestration
task.

Change-Id: I772b86d94b40b25e53344ec459942a5bf9b00b9f