flink chart: Add a comment label (2nd try)
While the previous patch, I772b86d94, did add a comment label it
ended up adding it at a level that was less useful. The goal was
to identify flinkdeployment releases, but the previous patch
ended up attaching it to the created pods and not the
flinkdeployment.

Move the comment label up to to the top level flinkdeployment
labels such that it will be queryable as expected.

Change-Id: I3d1ae3d459b2e47a203fe3801103ec70d419c67b