mediawiki: change the dump pods fsGroupChangePolicy to speed up volume mount
The dump pods have shown a gradual startup time slowdown in the last
couple of weeks. This is due to the fact that the volume contains more
and more files (currently ~1M) and that we chown _all_ files _at every
pod startup/volume mount_.

This is slow enough to hit a pod startup timeout in airflow and causing
the dump DAGs to massively fail.

> OnRootMismatch: Only change permissions and ownership if the permission
> and the ownership of root directory does not match with expected permissions
> of the volume. This could help shorten the time it takes to change ownership
> and permission of a volume.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T402644
Change-Id: I3ef73e0aa2c99b584654567c67be04612b2f675e