mediawiki: clean up php.version overrides
This change absorbs the common value of php.version that we set on all
mediawiki deployments ("8.1") into the chart and deletes all overrides.

Scap now sets php.version to ensure consistency with the selected image
flavour, so there's no value keeping the overrides (e.g., they will not
be reused / updated during the 8.3 migration).

The fact that the everywhere-overridden chart default was not advanced
to 8.1 at the end of the previous migration is simply an oversight.

Bug: T401721
Change-Id: If3c71fc6c3fda8d48a0b3a76d43edf91369cfed0