opensearch-cluster: Add chart for review (3/3)
This is the 3rd of 3 OpenSearch helm charts submitted for review.

The upstream chart can be found at
https://github.com/opensearch-project/opensearch-k8s-operator/tree/main/charts/opensearch-cluster

and we are using the 2.7.0 version of the chart.

This chart deploys OpenSearch cluster via the operator, which is installed in the
preceding chart.

See also the user guide at https://github.com/opensearch-project/opensearch-k8s-operator/blob/main/docs/userguide/main.md

Note that I have changed the scope of this patch, so previous patchsets have far more changes.
The current scope is limited to the chart itself, without any WMF-specific changes.

Bug: T397246
Change-Id: Ibb48b0c582e11d3b933cbded30386babb39bfb48