Facilitate a role swap between an-mariadb1001 and an-mariadb1002
We need to make an-mariadb1002 the active server temporarily, so that we
can upgrade the SSD firmware in an-mariadb1001.

This commit updates those applications that are deployed to Kubernetes
and currently refer to an-mariadb1001 to use the new hostname.

The change will be reverted on September 3rd, when the role swap will
also be reverted.

Bug: T394498
Change-Id: Ib341ae2cb01e297c0624b76993c6f1c46aaa517b