dse-k8s: Upgrade dse-k8s-codfw to v1.31 unpin charts
We are in the process of boostrapping the dse-k8s-codfw cluster with v1.31.
This patch ensures that the updated version uses the latest version of the charts available
and ensures that certmanager config reflects the changes to the chart. Also ensures
the leader election leases are created in the cert-manager namespace rather than kube-system.

Bug: T397301
Change-Id: I359412d2d591418392e096b316c09240d96bea45