changeprop-jobqueue: add CategoryCountUpdateJob rule
CategoryCountUpdateJob was recently added for T365303 and should be
isolated to its own rule in order to bound total concurrency.

The initial concurrency is 50, matching the limit it has been subject to
on the low-traffic consumer since its introduction (~ 1 week ago).

If changeprop concurrency metrics can be trusted, this is well above the
maximum observed concurrency during that time, which peaked at ~ 4 jobs.

As noted in the comment, we may need to introduce database section-based
partitioning in the future. An example of what that looks like is shown
in I042ac43d09fb1e07ad7eb8b8c25a1946abe93bc2 (not shown: manual Kafka
topic and partitioning setup steps).

Bug: T402873
Change-Id: Ie0f6f3d415085a2492adcc6e36abcbd2325c8a51