shellbox-syntaxhighlight: pilot 1 replica on 8.3
This change scales the migration release up to 1 replica, running the
first available image version with PHP 8.3.

In production, this will receive ~ 8% of traffic in both DCs.

Once this runs for a couple of hours without issue, the change will be
reverted and we'll move forward with preparation for larger migrations.

This includes aligning all shellbox services on the 2025-08-29-172844
image version, regardless of PHP version.

Bug: T403284
Change-Id: I143871b77ee7fe00d765a57da83fe2ec8764a4fe