ml-services: Disable autoscaling on edit-check model.
After the latest deployments configuring the batcher we still facing high latency on serving, so we decided to disable autoscaling and add a fix number of pods.
Set a fix number of 'minReplicas' which disables the autoscaling for edit-check model while we can keep the rest of the condfiguration same.
Changes on both staging/prod.

Bug: T403378
Change-Id: I1e5cb2fd74c37e817a7ba16e0a23e17ac6f800c9