dse-k8s: Augment the dse-k8s cluster namespaces.
We have been deploying new namespaces and config on the dse-k8s main key since we had only one dc.
However, we have recently deployed the dse-k8s-codfw cluster and need to separate the namespaces
and the accompanying configurations and limits per cluster with the main key holding the values for
shared config. This should be a NOOP for eqiad and should remove all the extra namespaces in codfw.

Bug: T404068
Change-Id: Ie758c9d26eaff1fb2f985eaed1b08bd2fb926437