shellbox-media: upgrade to PHP 8.3
Given the absence of issues seen on the previously migrated instances of
Shellbox, plus the fact that this is a comparatively lower-rps service,
it's unlikely that a prolonged fractional pilot would be valuable.

Instead, migrate shellbox-media in its entirety, which in practice will
be applied in a staggered manner across datacenters.

Bug: T403284
Change-Id: I74e5603a671f932b0007bfff4befe3be0c2412b6