flink-operator: allow upstream config to be applied
the upstream config present in the flink-operator chart got ignored
because of append: false.
The doc was particularly misleading since it suggested that you may want
this to allow overriding the config with your own default.
Here we want the upstream defaults because they contain valuable config
to make flink work with java17.

Consolidate WMF logger config back to the chart, I think jvm config
options from upstream are more important to use than keeping the
upstream logging config that we want to override anyways.

Bug: T404340
Change-Id: Ia59aa497e516c7660ef2497eb59bb04962fff4f4