mw-*: Tune 8.3 releases to prevent deployment timeouts
This revives temporary changes to maxUnavailable and helmfile operation
timeouts in Ia08474eeb156c9c6512c990008bac8928866ced2 for the migration
and next releases, ahead of the upcoming 8.3 production ramp.

In short, during the initial phase of the ramp, there is a period where
8.3 releases benefit little from reuse of cached image layers on the k8s
worker nodes - i.e., deployments slow down due to layers that are large,
but rare given the small initial footprint of these releases.

During the 8.1 migration, we found these particular settings were
effective in preventing deployment timeouts while the migration was in
progress.

More details / analysis in the commit description on that patch and in
https://phabricator.wikimedia.org/T383845#10612223.

Bug: T405955
Change-Id: I962163d84d014abfb29d3d4b471cd25d5ca5a43b