charts: add separate requests/limits for tegola's cronjob
The tegola's cronjob pods are responsible for tiles pregeneration,
and from a quick look at graphs it seems way heavier in resource
consumption than the main app pods.

Split the requests/limits into two categories, and double the limit
in production for the pregen pods.

This will allow to have a quicker cache warm-up time when needed.

Bug: T381565
Change-Id: I9eb2114c107a07c9d118a2843867071a4153c2e0