rest-gateway: use mw-api-ext rather than mw-api-int for all APIs
Historically, given that rest-gateway makes internal requests to the
APIs, we decided to use mw-api-int. However, this is an awkward
distinction, because the requests that the rest gateway makes are
externally generated. This was fine during the restbase migration, as
the few requests that went from restbase to mw-api *were*
internal. However, in reality those requests are now mostly directly
routed external requests. This was fine for a while.

However, now that we are rerouting rest.php (and soon action) APIs we
will incur a very significant increase in traffic volume. This would
cause a significant and unhealthy traffic swing from -ext to -int by
maintaining this pattern, and it needs to change. The time is now.

Our use of rest-gateway-ro at the edge via multi-dc.lua means that we
will route reads and writes appropriately to the active DC.

Add the ability to specify a per-endpoint stat prefix to better
distinguish restbase APIs from other mw-api-ext ones.

Bug: T401396
Change-Id: I2575f4f6e787dd9eede7952369f99cb10c0ae5d1