mw-debug: revert to default maxUnavailable
In Ia612b9752c11dd5686bfe5139e3c20e03f22b513, maxUnavailable and
maxSurge were set to 100% to speed up mw-debug deployments during the
scap testservers stage.

Unfortunately, with maxUnavailable: 100%, the helmfile apply calls will
return before the pods of the new replicaset are available, so when scap
runs httpbb checks, it's very likely hitting the *old* replicaset.

Scap does monitor deployment progress in terms of available replicas on
the new replicaset, but it's non-blocking - i.e., as soon as the apply
calls return, it's signaled to exit - so we can't really use that as a
mechanism to wait for readiness.

In any case, this reverts the maxUnavailable change, leaving only the
maxSurge: 100%. That's a solid optimization that we should keep, as it
allows the new pods to come up in parallel, rather than sequentially.

Change-Id: If0c883d4ebc4003c217c134c96e80ed901d39fff