opensearch-cluster: Add secrets and network policy templates to chart
Add a secret.yaml template, which is required for
setting a user/password combo other than "admin:admin".

See

https://github.com/opensearch-project/opensearch-k8s-operator/blob/v2.7.0/docs/userguide/main.md#custom-admin-user

for more details.

The idea is to use the secret template to create a K8s secret with a
valid username and password, which looks something like:

$ kubectl get secret opensearch-cluster-admin-password  -o yaml
apiVersion: v1
data:
  password: YWRtaW4=
  username: YWRtaW4=

(taken from minikube, note the default admin:admin)

After that, we reference the secret name as the
"adminCredentialsSecret" in the chart, something like:

  security:
    adminCredentialsSecret: opensearch-test-secret-config

Then we'll create a second secret we'll reference as the
"securityConfigSecret",
which is basically a template rendering of `internal_users.yml` as
described at
https://github.com/opensearch-project/opensearch-k8s-operator/blob/v2.7.0/docs/userguide/main.md#custom-admin-user

It requires a bcrypt-hashed version of the admin password, which is stored in the
first secret under the key 'hashed_password'.

Additionally, we create a Calico Network Policy which opens ports 9200 and 9300
(REST API and cluster state, respectively) to all pods with the `opster.io/opensearch-cluster`
label (read: all OpenSearch cluster nodes).

Bug: T397246
Change-Id: I98cb05f072653da89b905d4310dd7d5cbe1ead8e