spark-operator: Update RBAC rules for job namespaces
Previously, we used a technique of deploying an extra ClusterRole to the
namespaces tagged with `deployClusterRole: deploy-sparkapplications` in
order to allow the deploy user for that namespace to work with
sparkapplication and scheduledsparkapplication objects.

This technique is no longer necessary and it is not the way we work with
other operators, such as the flink, cloudnative-pg and opensearch
operators.

What we can do instead is to grant the permissions to the
`deploy` ClusterRole and assign it to the `deploy` RoleBinding.

This means that we can also remove the `deploy-sparkapplications`
ClusterRole and its configuration in the spark namespace.

Bug: T405490
Change-Id: I3785e5efbd0b08c703fa0ddf228cae4c7f403bec