Allow the spark-operator controller to communicate with driver pods
In recent versions of the spark-operator, work has begun to deprecate
the mutating webhook. The way it works now is that the controller
contacts the driver pod directly and adds a label:

`sparkoperator.k8s.io/mutated-by-spark-operator="true"`

Then the webhook only acts on pods with this label

In order to support this functionality, we need to enable a
networkpolicy that allows the spark-operator controller to communicate
with the driver pods in each of the job namespaces.

Bug: T405490
Change-Id: I0bdfcbaa268e587b233e32ba802eddbef11d05b2