mw-*: Right-size large service after switchover
This recalibrates replica counts for large services following the
September 2025 switchover, which is necessary to unblock capacity
exchanges between releases during the PHP 8.3 migration.

Multi-DC sizing is estimated using data accumulated so far following the
repool of eqiad on October 2. This targets a 60% FPM worker utilization
at the 0.95 quantile of observed load, taking the max over DCs, and also
adding a 10% bonus to compensate for the narrow time window.

Single-DC sizing is set based on a combination of:
1. actual replica counts used while serving purely from codfw and
2. estimates targeting the higher 75% at 0.95 of observed load we've
   used in the switchover (i.e., running hotter is part of the exercise).
The estimates in #2 are substituted in cases where the actual replica
counts in #1 were far higher than they needed to be. That's limited to
mw-api-int in this case.

Notably, this makes no changes to mw-api-ext, which simply maintains its
current single-DC sizing given the upcoming migration of rest-gateway
routed traffic from mw-api-int.

Replica count estimates were computed using the script in [0].

[0] https://gitlab.wikimedia.org/repos/sre/serviceops-kitchensink/-/merge_requests/24

Bug: T405955
Change-Id: I848254c26be994caeaa4b8746dd50440eacf8472