Add a deployment of the spark-support chart to our analytics-test namespace
We have created a new spark-support chart, which will manage the
resources required for spark-jobs to be useful.

This commit adds a helmfile configuration with a release of this chart
to the `analytics-test` namespace.

It also enables access to the analytics_test_hadoop
cluster, with suitable networkpolicies and configuration files.

Bug: T406833
Change-Id: I37f962e1fbac15ad9d85be9640c24b4142906c10