admin_ng: deploy the cluster role for the GPU node labeller to dse-k8s
The amd node labeller is a k8s controller that runs as daemon on
k8s workers equipped with a GPU. It reads /dev GPU metadata and
reports it to the k8s api.

This change adds the necessary clusterrole/binding support to make it
working.

The end result will be that a Deployment will be able to specify
what kind of GPU it is needed (VRAM, CU size, model, etc..), not
only that a generic one is requested.

Bug: T373806
Change-Id: I655378c13653023a2120fd3dc89290a3f0b5bb5f