airflow: allow the deployment of the triggerer component
This component allows tasks to be retried at a later date without
holding an entire pool slot in the meantime.

The ML team has expressed the need to deploy it to work around our GPU
being fully utilized, without crashing an entire DAG run.

The triggerer will be disabled by default and enabled as an override.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T406958
Change-Id: I8eb58c15d91b17dfe002327741d93d81c586b514