mw-(api-ext|web): Increase maxSurge on 8.3 "next" releases
The earlier changes in I962163d84d014abfb29d3d4b471cd25d5ca5a43b are
effective in avoiding deployment timeouts for the "next" releases.

However, they're still rather slow given that the image layers are
unlikely to be cached, typically taking 6-7 minutes.

One option to consider is increasing maxSurge (default is 3%) to allow
for some additional concurrency when creating new pods.

Given the small absolute size of these releases, setting this to a large
relative value is safe.

The proposed value of 10% is somewhat arbitrary, but intuitively this
allows us to achieve the same concurrency as the "migration" releases
where it's safe to use a larger maxUnavailable (20%).

Bug: T405955
Change-Id: I99267b3178106dbb43d0f73419160b4cc5755087